requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
distro==1.9.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-texttospeech==2.27.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy==2.2.6
openai==1.82.0
proto-plus==1.26.1
protobuf==5.29.4
pvporcupine==3.0.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
scipy==1.15.3
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
